Revised PR according to comments
